Sublime::Controller: stop search once a (tool) view is found
Sublime::Controller::eventFilter() looks for a view or a tool view
ancestor of a widget that is clicked on or receives focus. If such a
(tool) view ancestor is found, the search is stopped by returning early.
However, this search continues until the final main window ancestor in
case the found (tool) view ancestor matches the main window's active
(tool) view. A (tool) view cannot have another (tool) view as an
ancestor, so the further search is fruitless. Optimize by returning
early in this scenario too.