ItemViewWalker: clear previous selection
The actions "Jump to Next Outputmark" (F4) and
"Jump to Previous Outputmark" (Shift+F4) usually work well for the
Problems tool view: the first, next or previous row is selected and
activated, and the previously selected row (if any) is deselected.

However, when the user gives focus to the ProblemTreeView widget while
one of its rows is selected, then triggers one of the Jump actions, the
previously selected row is "pinned" and can no longer be deselected by
triggering the Jump actions for some reason (possibly due to a bug in
Qt 6.8.2). Once this "pinning" occurs, the first element of the list
returned from m_selectionModel->selectedRows() in
ItemViewWalker::selectIndex() is always the pinned index and the second
element - either the next or the previous index. These two indices are
selected in the UI even though the ProblemTreeView's selection mode
equals QAbstractItemView::SingleSelection.

As the intention is to replace the current selection rather than add to
it, replace the Current flag with the Clear flag in the selection
command in order to work around the bug.