Synchronize activeToolView with activeActionListener
When an IToolViewActionListener is raised or becomes visible during main
window state restoration, UiController assigns it to
UiControllerPrivate::activeActionListener but does not update
Sublime::MainWindowPrivate::activeToolView. If the activeToolView is
also an IToolViewActionListener and the user clicks on or otherwise
transfers focus to it, the signal
Sublime::MainWindow::activeToolViewChanged() is not emitted, because
nothing changes from the MainWindow's point of view. Therefore, the
user's attempts to activate an IToolViewActionListener surprisingly
fail.

Whenever a tool view that belongs to the active area is raised,
MainWindowPrivate::raiseToolView() is invoked (indirectly) from
UiController::raiseToolView(Sublime::View*). Replace updating the
activeActionListener in UiController::raiseToolView() with updating the
activeToolView in MainWindowPrivate::raiseToolView(). Changing the
activeToolView emits the signal activeToolViewChanged(), which lets the
UiController update the activeActionListener accordingly, and keeps the
two data members in sync. The activation of a raised tool view still
does not work perfectly as the added FIXME comment explains.

Manually update the activeToolView to the new value of the
activeActionListener when an IToolViewActionListener becomes visible
during main window state restoration. The relevant code is in the slot
UiController::toolViewVisibilityRestored() connected to the namesake
signal of Sublime::MainWindow. But the code cannot be moved to the
MainWindow, because only the UiController is aware of the interface
IToolViewActionListener and it needs to activate not just any of the
visible tool views but an IToolViewActionListener from among them.

Move the not-changed activeToolView check from
Sublime::Controller::eventFilter() to
Sublime::MainWindow::setActiveToolView() in order to let the added calls
to the latter function reuse it.