interfaces: disable copy&move of several abstract classes
Copying or moving instances of such classes is bug-prone, e.g. can cause
slicing. Such copying and moving is also almost never needed as
evidenced by the fact that disabling it does not break compilation. If
such operations are needed for some reason, the copy&move semantics
should be carefully and deliberately implemented or explicitly defaulted
instead of relying on the implementations implicitly generated by
compilers.