NativeAppJob: check for errors and early-return first
The values of m_name and m_killBeforeExecutingAgain are read from config
and discarded if an error is detected. Reorder code to not read them in
this case and thus optimize.

Make a config group object, that is only read from, const.

Do not needlessly explicitly specify the template argument of
KConfigGroup::readEntry().