Raise the relevant tool view when debug session ends
Whenever the current debug session ends (except on KDevelop exit),
DebugController::debuggerStateChanged() raises the Debug output tool
view in the Code sublime area. This works well in the most common case
when a debuggee runs and its output is shown in the Debug tool view. But
in less common scenarios described below, a debug job produces no
output, so raising the Debug tool view is pointless. It is also
confusing because the user can mistake the output of the previous debug
session for the output of the last debug session.

When the builder dependency job of a debug job fails or is canceled, the
debug job is never started and its output view is never added to the
Debug tool view. Raise the relevant Build output tool view in this
scenario.

When a debug job detects a launch configuration error, its dependency
job (if any) is never created and the debug job itself finishes without
adding an output view. Do not raise any tool view in this scenario.

Neither MIExamineCoreJob nor MIAttachProcessJob can have a dependency
job. These two jobs also do not inherit OutputJob and therefore cannot
show the output of the debuggee. Do not raise any tool view when the
debug session of such a job ends.

Always raise the Debug output tool view when a debug session other than
MIDebugSession ends in order to keep the behavior of external plugins
(such as kdev-python and kdev-xdebug) unchanged. The external plugins
can make use of the added flexibility by overriding
IDebugSession::toolViewToRaiseAtEnd().