Reorder and initialize DVcsJobPrivate's data members
Initialize the status and the vcsplugin data members of DVcsJobPrivate
instead of assigning to them in DVcsJob(). Keep creating the OutputModel
in DVcsJob() in order to immediately transfer the ownership to a base
class OutputJob and thus ensure that the model instance is not leaked.

Move the status data member beside the ignoreError data member in order
to reduce the size of DVcsJobPrivate by 8 bytes on a 64-bit system. Also
move the const pointer vcsplugin data member to a more logical place
higher up.