Pass Silent verbosity to VcsJob() from no-output jobs
3 bazaar job classes BzrAnnotateJob, CopyJob and DiffJob, and a git job
class StandardJob inherit class VcsJob, which in turn inherits class
OutputJob. But none of these 4 jobs outputs anything itself. The 4 jobs
delegate the actual work to helper jobs, some of which do produce
output. The constructors of these 4 job classes take an
OutputJobVerbosity parameter and forward it to VcsJob(). VcsJob()
invokes OutputJob::startOutput() in case of Verbose verbosity.

Out of these 4 jobs only the CopyJob ends up passing the Verbose
verbosity to VcsJob(), because BazaarPlugin::annotate() passes Silent to
BzrAnnotateJob(), the default value of DiffJob's verbosity parameter is
Silent, and Silent is passed to StandardJob() wherever an instance of
the class is created. A consequence of this verbosity of the CopyJob is
that when the user copies a bazaar-version-controlled file, an empty
output view appears in the Version Control output tool view before the
nonempty worker job's `bzr add` output view. The same useless empty
output view would be added for the other 3 job classes had they passed
the Verbose verbosity to VcsJob().

Always pass Silent verbosity from the constructors of the 4 jobs to
VcsJob(). Remove the no longer used verbosity parameter from 3 of the 4
job constructors. Keep the verbosity parameter of DiffJob() because it
is forwarded to the worker job that produces `bzr diff` output.