SvnJobBase: show svn progress dots on a separate line
As the addressed and removed TODO comment surmised, checking whether the
previous output line consists of dots works better than checking whether
it contains any dots.

The output of an svn job that commits 100 new files ended with
    Adding /path/to/file/98.txt.
    Adding /path/to/file/99.txt.....................................................................................................
before this commit and ends with
    Adding /path/to/file/98.txt.
    Adding /path/to/file/99.txt.
    ....................................................................................................
at this commit.

Similarly the output of an svn job that commits 100 modified files ended
with
    Sending /path/to/file/98.txt
    Sending /path/to/file/99.txt....................................................................................................
before this commit and ends with
    Sending /path/to/file/98.txt
    Sending /path/to/file/99.txt
    ....................................................................................................
at this commit.