subversion: show nonempty error message when job fails to start
When an svn job fails in start(), it calls
SvnJobBase::internalJobFailed() then KJob::setErrorText().
internalJobFailed() finds that internalJob()->errorMessage() is empty
and does not call setErrorText(). As a result, the job's error text ends
up empty while the error code equals 255 when it emits result, and
therefore RunController displays the empty error message in the
shell-embedded message area.

Extract SvnJobBase::failToStart() and call it instead of
internalJobFailed() and setErrorText() in the jobs' start() function in
order to enforce the correct order of calls. Set different error codes
when a job fails to start and when its internal job fails. Declare the
error codes in an unnamed enum as KJob's documentation recommends.

Do not try to retrieve internalJob()->errorMessage() in failToStart()
because it should always be empty when this function is called and
because the job already passes an appropriate error message as an
argument.

Do not try to output the error message in failToStart() because a job's
output is normally not started at this time. Outputting this error
message would not be useful either, because it is now displayed in the
shell-embedded message area. Creating an output view just to show the
same error message alone is pointless.

SvnCommitJob passes Verbose verbosity to its base class's constructor.
The Verbose verbosity is forwarded to VcsJob(), which therefore
asynchronously invokes OutputJob::startOutput(). SvnCommitJob::start()
also calls startOutput(). As a result, two output views are added to the
Version Control tool view and show identical svn commit output each time
the user makes a subversion commit in KDevelop.

Not calling startOutput() in SvnCommitJob::start() and relying on its
asynchronous invocation would not work, because start() is invoked
synchronously after the job's construction and immediately outputs
"Committing..." before the asynchronous invocation occurs.

Work around the SvnCommitJob's output view duplication by always passing
Silent to VcsJob(), then immediately setting the job's verbosity to
Verbose. Implement this workaround in SvnJobBase() and call
startOutput() for a Verbose job in SvnJobBase::startInternalJob(). This
change partially cancels the effect of
66b7a2b5b25dddb4b399878a8a897fad509c9dcd for svn jobs. I think the
"problem" referred to by the message of the aforementioned commit is
that calling startOutput() from VcsJob() occurs before derived job
classes have a chance to configure the OutputJob base class of VcsJob.
This problem does not apply to startInternalJob(), which is called from
the start() function of each derived job class.

Remove the dependency on the model type (QStandardItemModel) from
SvnCommitJob by adding a new optional string parameter named
introductoryOutputMessage to startInternalJob() and passing the message
"Committing..." to it.

Assert that model() is not null in SvnJobBase::outputMessage() instead
of checking and returning early if it is null. The assertion must not
fail now that SvnJobBase manages its output directly and alone.

Turn protected slots of SvnJobBase into private functions because they
are not and should not be called by derived classes and because they are
connected to only via the function pointer-based syntax.

Make a protected pointer SvnJobBase::m_part private and const.