VcsAnnotationItemDelegate: support empty commit message
An SVN commit message can be empty. QFontMetrics::horizontalAdvance()
returns 0 for an empty string. A QRect with width=0 has x2=x1-1 and
therefore is invalid. The messageRect is passed to
VcsAnnotationItemDelegate::doMessageLineLayout(), which asserts that it
is valid. So this assertion fails when the message of a commit that
appears in the Annotation is empty.

Set the width of messageRect to 1 if the commit message is empty to
prevent the assertion failure.