SvnJobBase: show all lines of a multiline message
SvnJobBase::outputMessage() creates a single QStandardItem for any
message. Therefore only one line of a multiline message is visible when
the Word Wrap option is off (it is off by default) because in this case
the uniformRowHeights property of the output tree view is true. For
example, only the second (middle) line of the following multiline error
message is visible by default:
    Error executing Job:
    Commit failed (details follow):
    Aborting commit: '/path/to/file' remains in conflict

Create a QStandardItem per message line so that a multiline message is
always visible in its entirety.