subversion: don't set output job title
Each svn job sets its object name to a specific translatable string in
the constructor. The object name of an output job is used as its output
tab title by default. Do not override the tab title with untranslatable
strings "Subversion" and "commit".

This change has no noticeable effect because all VCS jobs use the
Version Control standard output tool view, the type of which is
HistoryView that does not display tabs. But this might change in the
future, and the output title should be translatable as it is almost
everywhere else.