Deduplicate errorsFound() into KDevelop::makeVcsErrorJob()
PerforcePlugin::errorsFound() has most likely been copied from
GitPlugin::errorsFound(). The definitions of the two namesake functions
are almost identical except that the returned job types differ: DVcsJob
and GitJob. The differences between DVcsJob and its derived class GitJob
(UserType, GIT_OPTIONAL_LOCKS=0) are unimportant for the error job.

Extract a helper function that returns a DVcsJob to replace the two
duplicated functions and to be reused in the next commit.