QMakeJob: place output in Build rather than Run tool view
When the user triggers the action "Run QMake", a new output view is
added to the Run tool view with a long tab title equal to the name of
the QMakeJob: "Run QMake in /possibly/very/long/build/dir/path".
74bd22bc334b103d32eb7ffd71b9a731a4bcf684 introduced the line
`setStandardToolView( IOutputView::RunView );` without justification
while adding the QMakeJob class.

QMake is a tool akin to CMake that should run before Ninja/Make. Place
the QMake output into the Build tool view where it belongs, alongside
the CMake, Ninja and Make output. The type of the standard Build output
tool view is HistoryView, which does not display tabs, so the long tab
title issue is eliminated as well.