Build, run and skip VCS tests when required executable is missing
When the `bzr` executable is unavailable, test_kdevbazaar is never
built. Furthermore, BZR_FOUND=BZR_FOUND-NOTFOUND is written into the
CMake cache. So even when the `bzr` executable becomes available, the
test is not built until the BZR_FOUND variable is removed from the CMake
cache. And if the `bzr` executable is available during the initial CMake
configuration but then disappears, test_kdevbazaar is built, runs and
fails until the BZR_FOUND variable is removed from the CMake cache.

The same issues apply to the `git` executable and test_kdevgit. Nowhere
else in the KDevelop code base is find_program() used in this way.

Always build and run test_kdevbazaar and test_kdevgit. QSKIP() in the
tests' initTestCase() if the required executable is unavailable to
prevent misleading test failures or even crashes.

The QSKIP() idea is copied from test_lldb and test_lldbformatters, which
require the lldb-mi executable. Following the LLDB tests' example, skip
the VCS tests only at the end of initTestCase(). This way, creating a
BazaarPlugin/GitPlugin instance is always tested - that should not crash
or otherwise fail even if the required executable is unavailable.