test_kdevbazaar: don't set EXECUTABLE_OUTPUT_PATH
EXECUTABLE_OUTPUT_PATH is the old executable location CMake variable. It
is not set anywhere else in the KDevelop code base except in a manual
CMake test. Setting it does not affect the location of the
test_kdevbazaar executable (at least on my system) and is not needed for
the test to pass.