Don't lock DefinitionsRepo::mutex() recursively
Running `duchainify --dump-definitions <path>` invokes
Definitions::dump(). Definitions::dump() passes a DefinitionsVisitor to
DefinitionsRepo::visitAllItems() while holding the
DefinitionsRepo::mutex() lock. DefinitionsVisitor::operator() calls
Definitions::definitions(), which attempts to lock the same nonrecursive
mutex again and thus causes a deadlock.

This deadlock was introduced when the recursive mutex of the Definitions
repository was replaced with a nonrecursive one in
830a9aef9c1e1864512d839b21f2ff9a2762cbf4.

Move the definition of Definitions::definitions() into a new helper
function template definitionsInRepo() and generalize it. Pass the
already available DefinitionsRepo instance from DefinitionsVisitor to
definitionsInRepo() so as to reuse it instead of locking the mutex
again.

Placing definitionsInRepo() into an unnamed namespace and calling it in
DefinitionsVisitor::operator() requires moving the DefinitionsRepo
declaration and the ItemRepositoryFor<Definitions> specialization up.