Install kdevelop.notifyrc also in the unit test location
This prevents numerous warnings like the following while running GDB and
LLDB unit tests:
    kf.notifications: No event config could be found for event id
    "BreakpointHit" under notifyrc file for app "test_gdb"

The warning is printed since
https://commits.kde.org/knotifications/a3a17a4ef2a3c50450cd435f967e5338f1a7382f
if neither the user nor the global KNotifyConfig file references the
event ID of a sent notification. A global KNotifyConfig file is
installed for the KDevelop application but not for the unit tests.

Implement a helper function in AutoTestShell rather than in
KDevMI::Testing in order to let non-MI debugger tests, e.g.
connectiontest of kdev-xdebug, use it as well. Non-debugger tests may
also need to use the helper function if non-breakpoint notifications are
implemented in the future.

Make the new helper function separate from AutoTestShell::init() instead
of always initializing notifications in init(). This separation is
important because AutoTestShell::init() is also called from
non-unit-test programs, such as duchainify and testfiletemplates, but
the new helper function relies heavily on the macros like QVERIFY(),
which can only be used in a test function that is invoked by the test
framework (according to the macros' documentations).