debuggercommon: add QDebug support for MI::Value
When a breakpoint is updated, debug-print its row and the MI::Value
tuple that is applied to it.

I have noticed that the same tuple (formatted differently) can be found
in the "Debugger output" debug message only after implementing the
QDebug operators. Keep the operators for possible future use. Keep the
added debug message too because printing the row of the updated
breakpoint can be useful.