MIBreakpointController: set expression to "exp" field value
When the user creates a GDB watchpoint by clicking the "Stop on Change"
button in a variable tooltip, GDB/MI replies with a "wpt" tuple that
contains only two fields: "exp" and "number". The tuple is passed to
MIBreakpointController::updateFromDebugger(), which does not look for
the "exp" field, and therefore is unable to update the breakpoint's
location. As a result, the following debuggercommon warning is printed:
    Breakpoint doesn't contain required location/expression data
The same warning is printed by watchpoint test functions of test_gdb.

Look for the "exp" field and assign its value to the expression of the
updated breakpoint as a last resort before printing the warning. The
expression assignment is normally not useful, because the value of the
"exp" field equals the current expression value, but the warnings are
gone now.