Add xfce4-terminal item to external terminal combobox
The option --disable-server runs xfce4-terminal in a separate process,
independent from other xfce4-terminal processes. So no matter whether or
not another xfce4-terminal process is running, KDevelop considers the
executed program finished and outputs "*** Finished ***" only after the
xfce4-terminal window opened by KDevelop is closed, rather than
immediately while the child process is still running.

Another benefit of the --disable-server option is that the user can
reliably kill the child process and close the terminal window by
stopping the native app job in the KDevelop UI, and unrelated
xfce4-terminal processes started after the KDevelop's process are
unaffected by the stopping of the job. This behavior is consistent with
other external terminals - konsole and xterm.

The option --disable-server also works around
https://gitlab.xfce.org/apps/xfce4-terminal/-/issues/11
That is, passing this option ensures that KDevelop's Environment
configuration takes effect.

The option --hold keeps the xfce4-terminal window after the child
process exits, and therefore lets the user examine the process output.
KDevelop passes the corresponding options --hold and -hold to konsole
and xterm respectively.

The option -x, unlike -e, passes the contents of KDevelop's Arguments
line edit to the child process rather than to the xfce4-terminal
process, and thus prevents a failure to start with the following output
in case of e.g. Arguments="arg":
    /tmp> xfce4-terminal --disable-server --hold -e /path/to/app arg
    xfce4-terminal: Unknown option "arg"
    *** Failure: Exit code 1 ***

Place the xfce4-terminal item above the gnome-terminal item because the
support for gnome-terminal is impaired by difficult to fix bugs detailed
in the parent commit message.