Wait for exit of gnome-terminal external terminal
The benefit of the option --wait is that KDevelop considers the executed
program finished and outputs "*** Finished ***" only after the
gnome-terminal window is closed rather than immediately while the child
process is still running. This behavior is consistent with other
external terminals - konsole and xterm.

Unfortunately, when the user stops the native app job in the KDevelop
UI, the child process is not killed and the gnome-terminal window is not
closed. This differs from the behavior of the other supported external
terminals.

Another difficult to fix bug is that a gnome-terminal window is
automatically closed when its child process exits, which prevents the
user from examining the process output. Unlike the other supported
external terminal applications, gnome-terminal does not have a command
line option like --hold.

A frequently suggested workaround to the missing --hold option is to set
the command line to the following (adapted):
    gnome-terminal --wait -- sh -c '%exe;$SHELL'
But this command silently discards the contents of KDevelop's Arguments
line edit. And if KDevelop substitutes both the executable and its
arguments for the placeholder %exe as a workaround, the possible
resulting nested quotes can cause other bugs.