Use entire configured external terminal command for debugging
MIDebugSession::startDebugging() passes to STTY() only the name of the
executable out of the external terminal command configured on a native
launch configuration page. Then STTY appends hardcoded command line
options that depend on the external terminal application. Not only the
command line options for supported terminal applications are duplicated
in stty.cpp and nativeappconfig.ui, but options for an unsupported
terminal application configured by the user are ignored during
debugging.

The external terminal command constructed by STTY never exits on its
own, so the --hold command line option is not needed during debugging,
though it does not hurt either. The --disable-server option to
xfce4-terminal is useful during debugging because it allows KDevelop to
reliably close the terminal windows it creates. The --wait option to
gnome-terminal is not necessary during debugging, but without it there
is a risk that the gnome-terminal process exits too early and confuses
STTY, which does not expect such behavior from an external terminal.

Extract splitting of an external terminal command line into
ExecutePlugin::terminal(). Handle errors similarly to
ExecuteScriptPlugin::interpreter(). Return early from NativeAppJob() in
case of an error during the splitting, and reorder the splitting
invocation up so as to do less when such an error occurs.

Remove and deprecate the placeholder %exe, and append the executable and
its arguments to an external terminal command line. This ensures that
arguments are passed to the executable rather than to the external
terminal command. I have analyzed command line options of several
unsupported terminal applications, and found that those that are
compatible with the placeholder configuration invariably require the
placeholder to be placed at the end of the command line. So this change
should not substantially limit the number of compatible external
terminal applications.

Print a warning while removing a deprecated placeholder or deprecated
arguments to an external terminal command in order to prompt users to
clean up and update their external terminal configurations.

Move the list of supported external terminal command lines from
nativeappconfig.ui to a new helper function
IExecutePlugin::defaultExternalTerminalCommands() in order to let
gdb_test and lldb_test test the default command lines.

Connect to the signal STTY::OutOutput() in
MIDebugSession::startDebugging() unconditionally. STTY never emits this
signal when an external terminal is used, so the behavior is unchanged.
But the conditions under which this signal can be emitted is an
undocumented implementation detail of STTY, so MIDebugSession should not
depend on it by way of not connecting to the signal if an external
terminal is used. Optimize by reordering the connection to this signal
after the early-return branch that destroys the STTY object in
MIDebugSession::startDebugging()

STTY stores its external terminal QProcess in a QScopedPointer data
member. Therefore, the QProcess object is destroyed along with the STTY
and the MIDebugSession objects when the debug session ends. The
destruction of the QProcess closes the window of most supported external
terminal applications - konsole, xterm, and xfce4-terminal (unless the
terminal window has been transferred to another running xfce4-terminal
process). This is inconvenient because the user cannot read the final
output of the debugged program. This is also the source of an
inconsistency: when a program is run without debugging, the --hold
command line option is passed to external terminals, which ensures that
they are not automatically closed when the program exits.

Destroy STTY's external terminal QProcess object when the process exits
or on KDevelop exit, whichever comes first. Now the user is able to read
the entire output of the debugged program. Destroying the QProcess
objects on KDevelop exit serves multiple purposes:
1. KDevelop does not leave behind terminal windows after exit.
2. The external terminal windows, created in the test functions
   testDebugInExternalTerminal() of test_gdb and test_lldb, are closed
   when the test program exits. The annoyance of xfce4-terminal windows
   transferred to an unrelated running xfce4-terminal process and
   consequently left behind by the tests motivated this entire commit.
3. The QProcess objects are not leaked on KDevelop exit, and so will not
   be reported by tools that detect memory leaks.

When KDevelop or debugger unit tests exit, they still leave behind
gnome-terminal windows created by STTY, just as before this commit.
There is no reasonable way to close such windows because running the
gnome-terminal program unconditionally creates a separate process
gnome-terminal-server that is shared by all gnome-terminal windows.
KDevelop must not terminate the gnome-terminal-server process, because
doing so closes all gnome-terminal windows, including possible unrelated
ones opened manually by the user.

Apart from the debugging issue described in the previous paragraph,
non-debug execution in gnome-terminal also suffers from difficult to fix
bugs detailed in the message of a recent commit titled "Wait for exit of
gnome-terminal external terminal". Append a note to the tooltip of the
external terminal combobox to inform the user about the limited support
for gnome-terminal. Place the note in a separate translatable string so
as to be able to remove or tweak it without touching the text of the
main tooltip.

I do not use gnome-terminal, and so remove its package entirely in order
to prevent the annoyance while running unit tests. If it annoys another
KDevelop developer, the gnome-terminal data row can be skipped in
DebuggerTestBase::testDebugInExternalTerminal().

STTY::findExternalTTY() returns true if and only if the external
terminal process has started, even if an error that aborts the debug
session is detected. The return value is never used, so return void
instead of bool to simplify the code.