Don't include tag <br> in WatchpointHit notification text
The KDE Plasma notification server supports the tag <br> and displays a
line break in its place. But the Xfce notification server does not
support this tag (but supports e.g. the tag <b>). The behavior of the
Xfce server conforms to the Desktop Notifications Specification [1],
which recommends supporting the tags <b>, <i>, <u>, <a> and <img>, then
advises:
    A full-blown HTML implementation is not required of this spec, and
    notifications should never take advantage of tags that are not
    listed above.

So by including the tag <br> in the text of a WatchpointHit notification
KDevelop disregards the advice in the specification. As a result, these
tags are not even stripped and are displayed verbatim as "<br>" under
Xfce.

Use \n instead of <br> in the notification text in order to display the
intended line break no matter which notification server is running.

[1]:
https://specifications.freedesktop.org/notification-spec/latest-single/#markup