test_lldb: wait for state&idle instead of a fixed interval
LldbTest::testManualBreakpoint() consistently fails on my system (but
not on the FreeBSD CI server) with the following output:
FAIL!  : KDevMI::LLDB::LldbTest::testManualBreakpoint() Compared values are not the same
   Actual   (breakpoints()->rowCount()): 0
   Expected (1)                        : 1
   Loc: [kdevelop/plugins/lldb/unittests/test_lldb.cpp(722)]

Replace the calls to WAIT_FOR_A_WHILE() in
LldbTest::testManualBreakpoint() with the calls to
WAIT_FOR_STATE_AND_IDLE(). This makes the definitions of the slot
testManualBreakpoint() in LldbTest and GdbTest equivalent (sans
different MI command syntax and expected LLDB failures).

The session state is paused throughout the slot, so the test function
actually waits for the debugger to become idle. The debugger becomes
idle when all queued commands are sent and replied to, which is what
testManualBreakpoint() needs to wait for.
LldbTest::testManualBreakpoint() consistently passes on my system now.