debugger tests: don't disable breakpoint in the end
The loop in debugeeslow.cpp performs 3 iterations.
testChangeBreakpointWhileRunning() of test_gdb and test_lldb hits a
breakpoint during the first iteration, skips the second iteration while
the breakpoint is disabled and hits the enabled-again breakpoint during
the third iteration.

Do not disable the breakpoint once more during the supposed third
iteration to verify that the session ends without hitting the breakpoint
again, and consequently that all 3 iterations have occurred as expected.