test_lldb: revert several changes from test_gdb
Add and remove a breakpoint in 2 rather than 1 second after the debug
session becomes active to make an accidental stopping at the temporary
breakpoint and a consequent test failure less likely.

bd62debf695171f11576de26eba1a9436ff4f493 introduced a new test function
GdbTest::testInsertAndRemoveBreakpointWhileRunning() along with a fix
for Bug 201771. Judging by the following comment under the bug report,
the idea of the test is to add and quickly remove a breakpoint while the
debugee is running:
    I guess the reason why it only crashes while the debugee is running,
    is because then it takes longer to send the breakpoint to gdb, when
    the debugee is paused I can't click so fast to get it crash.
So waiting 200 milliseconds between the addition and the removal of a
breakpoint contradicts the test's purpose. The waiting also increases
the risk of an unintended stop at the longer-lived breakpoint and a
consequent test failure. Remove the waiting.

A side effect of not waiting for those 200 ms is that
LldbTest::testInsertAndRemoveBreakpointWhileRunning() prints the
following warning now:
QWARN  : kdevelop.plugins.debuggercommon: Received a modification of an unknown breakpoint
After printing the warning,
MIBreakpointController::notifyBreakpointModified() creates the already
removed breakpoint in the BreakpointModel again. This is acceptable
because the test does not care whether the breakpoint is registered in
the model, and the MI commands sent by KDevelop are unaffected by the
recreation of the breakpoint.
GdbTest::testInsertAndRemoveBreakpointWhileRunning() does not print this
warning, because only LLDB-MI sends two breakpoint-modified messages
when a breakpoint is added, even though the breakpoint is completely
unchanged.

Do not give debugeeslow extra time to run, because it should run less
than 7 seconds and WAIT_FOR_STATE() waits for 50 seconds.

Print the same debug message as
GdbTest::testInsertAndRemoveBreakpointWhileRunning() does before adding
a breakpoint for consistency.