Disable buttons when there are no analyzer check sets
The buttons "Clone check set" and "Remove check set" on the Clang-Tidy
and Clazy tabs of the Configure KDevelop dialog are always enabled when
the dialog is opened, even if there is no check set to clone or remove.
That's because when CheckSetSelectionListModel::reload() (invoked when
reset() is called on a Clang-Tidy or Clazy config page object during the
dialog initialization) resets the model, checkSetSelect emits the signal
QComboBox::currentIndexChanged() only if the model is not empty. The
calls to QComboBox::setCurrentIndex() in
CheckSetManageWidget::setCheckSetSelectionManager() and in
CheckSetManageWidget::reload() make checkSetSelect emit the signal
currentIndexChanged() only if the default check set index does not equal
the first available index, and both indices equal -1 when the model is
empty. Therefore, the function
CheckSetManageWidget::onSelectedCheckSetSelectionChanged(), which is
connected to the signal currentIndexChanged(), is not invoked when the
Configure KDevelop dialog is opened while the number of configured check
sets is zero.

Clicking the "Remove check set" button has no effect when there is no
check set to remove. However, clicking the "Clone check set" button,
entering a new check set name and pressing the OK button passes row=-1
to CheckSetSelectionListModel::cloneCheckSetSelection(), which causes
the following assertion failure:
ASSERT failure in QList::at: "index out of range", file /usr/include/qt6/QtCore/qlist.h, line 479

Assert that the current index is nonnegative in
CheckSetManageWidget::cloneSelectedCheckSetSelection() in order to fail
earlier and sort of document this precondition. Now if the added
assertion fails, it informs the developer that the function must not be
invoked, and therefore the clone button must be disabled when the
current check set index is invalid.

Instead of setting the current index of checkSetSelect after connecting
onSelectedCheckSetSelectionChanged() to its signal
currentIndexChanged(), set it before the connection and call the
connected-later function manually. Now both the clone button and the
remove button are enabled if and only if at least one check set is
configured.