Disable the rename button when there are no analyzer check sets
When there are no check sets, the clone, remove and set-as-default
buttons are disabled on the Clang-Tidy and Clazy tabs of the Configure
KDevelop dialog. But the rename button is enabled for some reason, even
though there is no check set to rename. Entering a new check set name
and pressing the OK button has no effect then. Disable the useless and
possibly misleading rename button as well in this scenario.