Don't reset Analyzer and Runtime config pages during initialization
Not doing the extra work improves performance. The UI of each of the 6
config page classes optimized in this commit displays the current config
values even if reset() is not called during initialization. I do not
quite understand why calling reset() is not needed, because these page
classes have non-null ConfigPagePrivate::configManager, and
KDevelop::ConfigPage::reset() calls
KConfigDialogManager::updateWidgets(), the documentation of which
contains the following sentence:
    Example use: Initialisation of dialog.
Perhaps this sentence is incorrect or obsolete. Or maybe updateWidgets()
needs to be called only in some scenarios.

This commit gets rid of the following warning when the Configure
KDevelop dialog is opened while there is no stored Clang-Tidy config:
    kdevplatform.shell: Settings in config page
    ClangTidyPreferences(<pointer>, name = "ClangTidyPreferences")
    changed, while UiPreferences(<pointer>, name="UiConfig") is
    currently selected. This case is not implemented yet.
The two counterpart warnings for AndroidPreferences and
Heaptrack::GlobalConfigPage in place of ClangTidyPreferences are also
eliminated.

This commit partially addresses the TODO comment in
KDevelop::ConfigPage::needsResetDuringInitialization(). The added
functions are inline because their definitions are not going to change
and because declaring and defining each function in a single place will
make its eventual removal (envisioned in the aforementioned TODO
comment) easier.